# Reduce cpu, memory, and allocations in formatter